Split axis settings into separate modules
